@@ -80,10 +80,10 @@
     "jsdom": "~22.1.0",
     "jsonc-eslint-parser": "^2.1.0",
     "ng-packagr": "^17.3.0",
-    "prettier": "^2.6.2",
+    "prettier": "3.1.1",
     "ts-jest": "^29.1.0",
     "ts-node": "10.9.1",
-    "typescript": "^5.0.0",
+    "typescript": "5.2.2",
     "vite": "^5.0.0",
     "vitest": "^1.0.4"
   },
