@@ -1,23 +1,12 @@
-import { ApplicationConfig } from '@angular/core';
-import { provideRouter } from '@angular/router';
-import { provideHttpClient, withInterceptorsFromDi, HTTP_INTERCEPTORS } from '@angular/common/http';
-import { appRoutes } from './app.routes';
-import { AuthInterceptor } from '@enterprise-dashboard/core';
-import { ContentTypeInterceptor } from '@enterprise-dashboard/core';
+import { ApplicationConfig } from "@angular/core";
+import { provideRouter } from "@angular/router";
+import {
+  provideHttpClient,
+  HTTP_INTERCEPTORS,
+  withInterceptorsFromDi,
+} from "@angular/common/http";
+import { appRoutes } from "./app.routes";
 
 export const appConfig: ApplicationConfig = {
-  providers: [
-    provideRouter(appRoutes),
-    provideHttpClient(withInterceptorsFromDi()),
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: AuthInterceptor,
-      multi: true
-    },
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: ContentTypeInterceptor,
-      multi: true
-    }
-  ],
+  providers: [provideRouter(appRoutes), provideHttpClient()],
 };
