@@ -1,16 +1,18 @@
-import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators, ReactiveFormsModule } from '@angular/forms';
-import { Router } from '@angular/router';
-import { CommonModule } from '@angular/common';
-import { ApiService, LoginRequest } from '../../services/api.service';
+import { Component, OnInit } from "@angular/core";
+import {
+  FormBuilder,
+  FormGroup,
+  Validators,
+  ReactiveFormsModule,
+} from "@angular/forms";
+import { Router } from "@angular/router";
+import { CommonModule } from "@angular/common";
+import { ApiService, LoginRequest } from "../../../services/api.service";
 
 @Component({
-  selector: 'app-login',
+  selector: "app-login",
   standalone: true,
-  imports: [
-    CommonModule,
-    ReactiveFormsModule
-  ],
+  imports: [CommonModule, ReactiveFormsModule],
   template: `
     <div class="login-container">
       <div class="login-card">
@@ -19,38 +21,45 @@ import { ApiService, LoginRequest } from '../../services/api.service';
           <p>Sign in to your account</p>
         </div>
 
-        <form [formGroup]="loginForm" (ngSubmit)="onSubmit()" class="login-form">
+        <form
+          [formGroup]="loginForm"
+          (ngSubmit)="onSubmit()"
+          class="login-form"
+        >
           <div class="form-group">
             <label for="email">Email/Username</label>
-            <input 
+            <input
               [formControlName]="email"
-              id="email" 
-              type="text" 
+              id="email"
+              type="text"
               class="form-input"
               placeholder="admin"
-              [disabled]="isLoading">
+              [disabled]="isLoading"
+            />
             <div class="error-hint">Use "admin" for username</div>
           </div>
 
           <div class="form-group">
             <label for="password">Password</label>
-            <input 
+            <input
               [formControlName]="password"
-              id="password" 
-              type="password" 
+              id="password"
+              type="password"
               class="form-input"
               placeholder="Enterprise123!"
-              [disabled]="isLoading">
+              [disabled]="isLoading"
+            />
           </div>
 
           <div *ngIf="errorMessage" class="error-message">
             {{ errorMessage }}
           </div>
 
-          <button 
-            type="submit" 
+          <button
+            type="submit"
             class="login-btn"
-            [disabled]="loginForm.invalid || isLoading">
+            [disabled]="loginForm.invalid || isLoading"
+          >
             <span *ngIf="!isLoading">Sign In</span>
             <span *ngIf="isLoading">Signing in...</span>
           </button>
@@ -63,28 +72,30 @@ import { ApiService, LoginRequest } from '../../services/api.service';
       </div>
     </div>
   `,
-  styleUrls: ['./login.component.css']
+  styleUrls: ["./login.component.css"],
 })
 export class LoginComponent implements OnInit {
   loginForm: FormGroup;
   isLoading = false;
-  errorMessage = '';
+  errorMessage = "";
+  email = "";
+  password = "";
 
   constructor(
     private fb: FormBuilder,
     private apiService: ApiService,
-    private router: Router
+    private router: Router,
   ) {
     this.loginForm = this.fb.group({
-      email: ['', [Validators.required, Validators.email]],
-      password: ['', Validators.required]
+      email: ["", [Validators.required, Validators.email]],
+      password: ["", Validators.required],
     });
   }
 
   ngOnInit() {
     // Redirect if already authenticated
     if (this.apiService.isAuthenticated()) {
-      this.router.navigate(['/dashboard']);
+      this.router.navigate(["/dashboard"]);
     }
   }
 
@@ -94,25 +105,27 @@ export class LoginComponent implements OnInit {
     }
 
     this.isLoading = true;
-    this.errorMessage = '';
+    this.errorMessage = "";
 
     const credentials: LoginRequest = {
       email: this.loginForm.value.email,
-      password: this.loginForm.value.password
+      password: this.loginForm.value.password,
     };
 
     this.apiService.login(credentials).subscribe({
       next: () => {
-        const returnUrl = this.router.parseUrl(this.router.url).queryParams['returnUrl'] || '/dashboard';
+        const returnUrl =
+          this.router.parseUrl(this.router.url).queryParams["returnUrl"] ||
+          "/dashboard";
         this.router.navigate([returnUrl]);
       },
       error: (error: any) => {
-        this.errorMessage = error.message || 'Login failed. Please try again.';
+        this.errorMessage = error.message || "Login failed. Please try again.";
         this.isLoading = false;
       },
       complete: () => {
         this.isLoading = false;
-      }
+      },
     });
   }
 }
